popcnt for Go 1.9
